+++ aplikacjaselenium/pom.xml
@@ -52,7 +52,7 @@
             <artifactId>selenium-java</artifactId>
             <version>3.141.59</version>
 		</dependency>
-		
+
 	</dependencies>
 
 	<build>
@@ -61,6 +61,19 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
+			<plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.19.1</version>
+                <dependencies>
+                    <!-- let surefire run JUnit 4 but also JUnit 5 tests -->
+                    <dependency>
+                        <groupId>org.junit.platform</groupId>
+                        <artifactId>junit-platform-surefire-provider</artifactId>
+                        <version>1.1.0</version>
+                    </dependency>
+                </dependencies>
+            </plugin>
 		</plugins>
 	</build>
