+++ aplikacjaselenium/src/test/java/pl/pandait/panda/PandaApplicationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PandaApplicationTests {
+class PandaApplicationTest {
 
 	@Test
 	void contextLoads() {
